Big change Made a ProfileCheck component that will render either CreateProfile or ShowProfile component based on if there is a profile id.
There are problems setting the state of profileId and the page is rendering
the CreateProfile component before the .then async actions end.

Ultimately I believe the problem lies within the ShowProfile component. Somehow
the user is not defined and I think this is because of it's former relationship
with App.js is tied to ProfileCheck.js and somewhere the user prop is not sent down.